+++ documentation/diagrams.md
@@ -1,3 +1,1 @@
 Edit & render diagrams at https://excalidraw.com
-
-API Sketch: https://excalidraw.com/#json=D-SBDD8b-I55-G6RI4Gpi,cShz70r5orDZFgNnjIwPKA
