--- src/main/java/ro/unibuc/contact/config/AppConfig.java
@@ -1,14 +1,0 @@
-package ro.unibuc.contact.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class AppConfig {
-
-    @Bean
-    public RestTemplate getRestTemplate() {
-        return new RestTemplate();
-    }
-}
