--- src/main/java/ro/unibuc/contact/data/MessageRepository.java
@@ -1,17 +1,0 @@
-package ro.unibuc.contact.data;
-
-import java.util.List;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- * No need to implement this interface.
- * Spring Data MongoDB automatically creates a class it implementing the interface when you run the application.
- */
-@Repository
-public interface MessageRepository extends MongoRepository<MessageEntity, String> {
-
-    List<MessageEntity> findByUserId(String userId);
-
-}
