--- src/main/java/ro/unibuc/contact/data/UserRepository.java
@@ -1,22 +1,0 @@
-package ro.unibuc.contact.data;
-
-import java.util.Optional;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- * No need to implement this interface.
- * Spring Data MongoDB automatically creates a class it implementing the interface when you run the application.
- */
-@Repository
-public interface UserRepository extends MongoRepository<UserEntity, String> {
-
-    Optional<UserEntity> findByEmail(String email);
-    Optional<UserEntity> findByUsername(String username);
-    Optional<UserEntity> findById(String id);
-    void deleteById(String id);
-
-    Boolean existsByUsername(String username);
-
-}
