--- src/main/java/ro/unibuc/contact/dto/Greeting.java
@@ -1,32 +1,0 @@
-package ro.unibuc.contact.dto;
-
-public class Greeting {
-
-    private long id;
-    private String content;
-
-    public Greeting() {
-    }
-
-    public Greeting(long id, String content) {
-        this.id = id;
-        this.content = content;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-}
