--- src/main/java/ro/unibuc/contact/dto/UpdateUserPasswordRequest.java
@@ -1,20 +1,0 @@
-package ro.unibuc.contact.dto;
-
-public class UpdateUserPasswordRequest {
-    private String password;
-
-    UpdateUserPasswordRequest() {
-        
-    }
-    UpdateUserPasswordRequest(String password) {
-        this.password = password;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
