--- src/main/java/ro/unibuc/contact/exception/EntityNotFoundException.java
@@ -1,10 +1,0 @@
-package ro.unibuc.contact.exception;
-
-public class EntityNotFoundException extends RuntimeException {
-
-    private static final String entityNotFoundTemplate = "Entity: %s was not found";
-
-    public EntityNotFoundException(String entity) {
-        super(String.format(entityNotFoundTemplate, entity));
-    }
-}
