+++ src/main/java/ro/unibuc/hello/config/AppConfig.java
@@ -1,0 +1,14 @@
+package ro.unibuc.hello.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class AppConfig {
+
+    @Bean
+    public RestTemplate getRestTemplate() {
+        return new RestTemplate();
+    }
+}
