+++ src/main/java/ro/unibuc/hello/data/MongoConfig.java
@@ -1,0 +1,39 @@
+package ro.unibuc.hello.data;
+
+import com.mongodb.ConnectionString;
+import com.mongodb.MongoClientSettings;
+import com.mongodb.client.MongoClient;
+import com.mongodb.client.MongoClients;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.mongodb.config.AbstractMongoClientConfiguration;
+
+import java.util.Collection;
+import java.util.Collections;
+
+@Configuration
+public class MongoConfig extends AbstractMongoClientConfiguration {
+
+    @Value("${mongodb.connection.url}")
+    private String connectionURL;
+
+    @Override
+    protected String getDatabaseName() {
+        return "test";
+    }
+
+    @Override
+    public MongoClient mongoClient() {
+        MongoClientSettings mongoClientSettings = MongoClientSettings.builder()
+                .applyConnectionString(new ConnectionString(connectionURL))
+                .build();
+
+        return MongoClients.create(mongoClientSettings);
+    }
+
+    @Override
+    public Collection getMappingBasePackages() {
+        return Collections.singleton("ro.unibuc.hello.data");
+    }
+}
