+++ src/main/java/ro/unibuc/hello/exception/EntityNotFoundException.java
@@ -1,0 +1,10 @@
+package ro.unibuc.hello.exception;
+
+public class EntityNotFoundException extends RuntimeException {
+
+    private static final String entityNotFoundTemplate = "Entity: %s was not found";
+
+    public EntityNotFoundException(String entity) {
+        super(String.format(entityNotFoundTemplate, entity));
+    }
+}
