--- src/test/java/ro/unibuc/contact/service/HelloWorldServiceTestIT.java
@@ -1,33 +1,0 @@
-package ro.unibuc.contact.service;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import ro.unibuc.contact.data.InformationRepository;
-import ro.unibuc.contact.dto.Greeting;
-
-@SpringBootTest
-@Tag("IT")
-class HelloWorldServiceTestIT {
-
-    @Autowired
-    InformationRepository informationRepository;
-
-    @Autowired
-    HelloWorldService helloWorldService;
-
-    @Test
-    void test_buildGreetingFromInfo_returnsGreetingWithInformation() {
-        // Arrange
-        String title = "Overview";
-
-        // Act
-        Greeting greeting = helloWorldService.buildGreetingFromInfo(title);
-
-        // Assert
-        Assertions.assertEquals(1, greeting.getId());
-        Assertions.assertEquals("Overview : This is an example of using a data storage engine running separately from our applications server!", greeting.getContent());
-    }
-}
